Move boot the extension helper
